+++ .vitepress/theme/components/Category.vue
@@ -29,4 +29,7 @@ const data = computed(() => initCategory(theme.value.posts))
     color: var(--bt-theme-title);
     font-family: var(--date-font-family),serif;
 }
+.posts{
+    text-decoration: none;
+}
 </style>